refactor: 拆分 temporary image URL 偵測與佔位符邏輯至獨立模組
- 將 isTemporaryImageUrl 函數從 imageUtils.js 拆分至 temporaryImageUrl.js，以避免在 background bundle 中引入整個 imageUtils 物件，減少 bundle 大小。
- 將 buildTemporaryImagePlaceholderBlock 函數移至 temporaryImagePlaceholder.js，確保其僅在 content side 使用，並避免不必要的打包。
- 更新相關測試以反映模組結構的變更，確保功能正常運作。
